#116: address review-reuse — extract NsDataExt, fix Keychain attrs and errors
- toNSData()/toByteArray() extracted to appleMain/util/NsDataExt.kt;
  Keychain and MdnsDiscovery now share the helper instead of carrying
  a duplicate inline.
- kSecAttrAccessible moved from kSecPrivateKeyAttrs to the top-level
  attrs dict, where SecKey.h requires it; it was silently ignored
  where it lived before, weakening the AfterFirstUnlock guarantee.
- errSecNotAvailable error message now describes a keychain-service
  outage (which is what the code actually means) instead of an
  entitlement issue; errSecMissingEntitlement gets its own branch
  with the entitlement-specific message.
- publicKeyFromRawPoint deleted from Keychain.apple.kt — it had no
  production caller; the test now carries its own file-local
  parseP256PublicKey helper.
