+++ composeApp/src/appleMain/kotlin/com/tubetoast/tether/util/NsDataExt.kt
@@ -1,0 +1,16 @@
+@file:OptIn(ExperimentalForeignApi::class)
+
+package com.tubetoast.tether.util
+
+import kotlinx.cinterop.ExperimentalForeignApi
+import kotlinx.cinterop.addressOf
+import kotlinx.cinterop.usePinned
+import platform.Foundation.NSData
+import platform.Foundation.dataWithBytes
+
+internal fun ByteArray.toNSData(): NSData? {
+    if (isEmpty()) return NSData()
+    return usePinned { pinned ->
+        NSData.dataWithBytes(pinned.addressOf(0), size.toULong())
+    }
+}
